Terrible bugfix for a known problem with submodels
Basically, since we delegate all operations to the main model, the main model
needs to know the root of the submodel for the purposes of `update`. This is
terrible and means there is an implementation issue. For now, it's fine because
we code against `protocols.Model[State]` and not the specific types themselves
but I am sure it will come back to bite someone.