Update node ops while iterating
As an artifact of the previous optimization in
a65fe51dec0c85f4f2e004b18397516fa5e7b821, a necessary side effect was lost. This
wouldn't really impact anything externally as anything that was observed would
still use the previous __get_node_for_ops function.

The best thing to do would be to centralize the iteration again, but use a
generator to implement the iteration. However, as time does not currently permit
this, a simple fix is sufficient.